Add user creation functionality
The HomeController, WishService, and User model have been updated to add functionality for creating new users. A form on the start page has also been updated to include a user input field and a submit button to add a new user. The create user function in the WishService now calls the database as opposed to hardcoded data. The add wishlist function has been moved but no changes have been made to it.