Update user handling and CSS path in wishlist application
Reworked the method of getting users and their wishlists in HomeController and WishService. The CSS file path in various HTML templates was moved from /templates/ to the project root. The parameter order in 'addWishlist' method was also switched for better data handling. User identification is now more accurate as well with use of userID over username for data queries.