Refactor variable names and enhance user management in HomeController
Detailed refactor of the 'Wishlist' and 'User' models to improve the clarity of variable names. In 'Wishlist', 'wishlistName' was changed to 'wishlist_name'. In 'User', 'userName' was changed to 'username', and 'passWord' was changed to 'user_password'. A user is now identified by 'userID' instead of 'id' in 'HomeController'. This change improves the precision of user identification and enhances the readability of the code base.