Diagrams added: ClassDiagramNew.png, ER_Diagram_New.png, Sequence_Diagram_New.png.
Optimize naming for fields and attributes

The naming of fields in the database and their corresponding objects in User, Item and Wishlist classes has been refactored for enhanced clarity. The revised naming also applies to their references in SQL scripts, Java classes, and HTML templates, ensuring consistency throughout the code.