Update login process and style enhancements
Refactored the login process in HomeController to identify users by email instead of username and made corresponding form updates in the loginpage.html. Updated the MySQL database password key in application.properties and added loggedInUser to the model in wishlist page. Moreover, background gradient style is added in style.css for visual enhancement.